Move main components into containers.
