+++ client/app/bundles/Events/containers/GroupDetail.jsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import { connect } from 'react-redux';
-import UpcomingEvent from './UpcomingEvent';
+
+import UpcomingEvent from '../components/UpcomingEvent';
 import { fetchGroup } from '../actions';
 
 class GroupDetail extends Component {
