+++ client/app/bundles/Events/routes/routes.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { Route, Switch } from 'react-router';
 
 import Events from '../containers/Events';
-import EventDetail from '../components/EventDetail';
-import Attendances from '../components/Attendances';
+import EventDetail from '../containers/EventDetail';
+import Attendances from '../containers/Attendances';
 import Groups from '../containers/Groups';
 import Members from '../containers/Members';
-import GroupDetail from '../components/GroupDetail';
+import GroupDetail from '../containers/GroupDetail';
 import Profile from '../containers/Profile';
-import MemberDetail from '../components/MemberDetail';
-import EventImport from '../components/EventImport';
+import MemberDetail from '../containers/MemberDetail';
+import EventImport from '../containers/EventImport';
 import AttendanceMatching from '../containers/AttendanceMatching';
 import AttendanceImport from '../containers/AttendanceImport';
 import Dashboard from '../containers/Dashboard';
